@@ -17,24 +17,24 @@
 	"devDependencies": {
 		"@playwright/test": "^1.45.1",
 		"@sveltejs/adapter-auto": "^3.2.2",
-		"@sveltejs/kit": "^2.5.18",
-		"@sveltejs/vite-plugin-svelte": "^3.1.1",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@typescript-eslint/eslint-plugin": "^7.16.1",
 		"@typescript-eslint/parser": "^7.16.1",
 		"autoprefixer": "^10.4.19",
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.42.0",
+		"eslint-plugin-svelte": "^2.45.1",
 		"postcss": "^8.4.39",
 		"postcss-load-config": "^5.1.0",
 		"prettier": "^3.3.3",
 		"prettier-plugin-tailwindcss": "^0.5.14",
-		"svelte": "^4.2.18",
-		"svelte-check": "^3.8.4",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
 		"tailwindcss": "^3.4.5",
 		"tslib": "^2.6.3",
 		"typescript": "^5.5.3",
-		"vite": "^5.3.3",
+		"vite": "^5.4.4",
 		"vitest": "^1.6.0"
 	},
 	"type": "module",
