@@ -1,5 +1,9 @@
 <script lang="ts">
-    export let toastMessage: string = "failed"
+    interface Props {
+        toastMessage?: string;
+    }
+
+    let { toastMessage = "failed" }: Props = $props();
 </script>
 
 <div id="toast-danger" class="flex items-center w-full max-w-xs p-4 mb-4 text-gray-500 bg-white rounded-lg shadow dark:text-gray-400 dark:bg-gray-800 mx-auto" role="alert">
