@@ -1,5 +1,7 @@
-<script>
+<script lang="ts">
 	import '../app.pcss';
+
+	let { children } = $props();
 </script>
 
-<slot></slot>
+{@render children()}
