@@ -1,37 +1,46 @@
 <script lang="ts">
 	import type { PageData } from './$types';
 	import { onMount } from 'svelte';
-	import type { FuelRefillDatum } from './+page.ts';
 	import { goto } from '$app/navigation';
 	import ToastFailed from '$lib/components/toastFailed.svelte';
 	import ToastSucceeded from '$lib/components/toastSucceeded.svelte';
-	export let data: PageData;
+	interface Props {
+		data: PageData;
+	}
+
+	let { data }: Props = $props();
 
 	onMount(async () => {
 		const { initFlowbite } = await import('flowbite');
 		initFlowbite();
 	});
 
-	$: to = Math.min(data.pageIndex * data.pageSize, data.fuelRefill.totalRecord);
+	let to = $derived(Math.min(data.pageIndex * data.pageSize, data.fuelRefill.totalRecord));
 
-	$: showing = (data.pageIndex - 1) * data.pageSize + 1;
+	let showing = $derived((data.pageIndex - 1) * data.pageSize + 1);
 
-	let fuelRefillData: FuelRefillDatum[] = [];
-	$: if (data.fuelRefill.fuelRefillData) {
-		fuelRefillData = data.fuelRefill.fuelRefillData;
-	}
+	let fuelRefillData = $derived.by(() => {
+		if (data.fuelRefill.fuelRefillData) {
+			return data.fuelRefill.fuelRefillData;
+		}
+		return [];
+	});
 
-	$: nextPage = data.pageIndex + 1;
-	$: if (data.pageIndex >= data.fuelRefill.totalPage) {
-		nextPage = data.fuelRefill.totalPage;
-	}
+	let nextPage = $derived.by(() => {
+		if (data.pageIndex >= data.fuelRefill.totalPage) {
+			return data.fuelRefill.totalPage;
+		}
+		return data.pageIndex + 1;
+	});
 
-	$: previousPage = data.pageIndex - 1;
-	$: if (data.pageIndex <= 1) {
-		previousPage = 1;
-	}
+	let previousPage = $derived.by(() => {
+		if (data.pageIndex <= 1) {
+			return 1;
+		}
+		return data.pageIndex - 1;
+	});
 
-	let toastMessage = ''; // case noStatus
+	let toastMessage = $state(''); // case noStatus
 	switch (data.fuelRefill.showToast) {
 		case 'noStatus':
 			toastMessage = '';
@@ -128,7 +137,7 @@
 					<tbody>
 						{#each fuelRefillData as fuelRefillDatum}
 							<tr
-								on:click={() => {
+								onclick={() => {
 									goto(
 										`/fuel/refills/${fuelRefillDatum.id}?currentUserId=${data.currentUser?.id}&currentCarId=${data.currentCar?.id}`
 									);
